requirements.txt
openai==1.63.2
pydantic==2.10.6
pytest==8.3.4
colorama==0.4.6
python-dotenv==1.0.1
google-generativeai>=0.8.3
